blog/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
blog==1.0.1
cached-property==1.5.2
celery
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.20
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser
docopt==0.6.2
exceptiongroup==1.1.2
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu
Markdown
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.1
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
random-password-generator==2.2.0
random-username==1.0.2
Random-Word-Generator==1.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
